improve the logic and view of the score
